error: include backtrace for errors originating from rust
The WebDriver standard says the stacktrace field on error responses are
implementation-defined, but the field cannot be omitted.  The errors
originating from Marionette are passed on without modification and
include the expected stacktrace.

This exposes a backtrace from the Rust crate with the same name for the
errors originating from within the library itself.

Fixes: https://github.com/mozilla/webdriver-rust/issues/63