command, response: align timeouts format with spec
WebDriver's timeouts configuration object format changed recently,
and this aligns us with the specification.  Fields in the JSON Object
passed to Set Timeouts as parameters are optional, meaning it is possible
to send {"search": <u64>} without defining "pageLoad" or "implicit".
Upon responding, the JSON Object in the body must contain all fields.