server: improve logging of http requests and responses
HTTP requests and responses will now be logged in a
similar style as Marionette chatter in geckodriver (following
https://github.com/mozilla/geckodriver/pull/636) and in the Marionette
server in Gecko, by employing arrows.

The logging of responses is reduced from two log entries, to one.
It also makes use of the fmt::Display trait of the status.

For requests, the HTTP method, URI, and full body is included.