response: correct window and element rect types
It was agreed at the WebDriver WG work day on Wednesday
that the appropriate return type for ElementRectResponse and
WindowRectResponse would be signed 32-bit integers because the WebIDL
for screenY/screenX/outerWidth/outerHeight mark them as longs.

This means that whilst it is technically possible to set a window to a
decimal-point value, the serialised return value will always be a long.